requirements.txt
aiogram>=3.15
python-dotenv>=1.0
aiosqlite>=0.20
